Extract shared primitives and rename MerchantRow to Merchant
Pulls pure helpers (validators, regex, SQL CASE builders, source-priority
ladder, did_you_mean) and Pydantic input models into a new
categorization/_shared.py. Drops the leading underscore on
SOURCE_PRIORITY, priority_case_sql, match_shape_case_sql, and did_you_mean
now that they cross module boundaries — the underscore was honest when
they were file-private and would be misleading at the package boundary.

Renames MerchantRow to Merchant: the type carries identity + recognition
rules + default mapping, all attributes of one entity in the user's
merchant catalog (app.user_merchants), not a transient match result.
Stays a NamedTuple to preserve positional unpacking.

No behavior change; 2058 tests pass.
